+++ R/taxize-package.R
@@ -49,7 +49,6 @@
 #' @importFrom utils URLencode citation download.file read.delim write.table
 #' @importFrom ape read.tree as.phylo.hclust plot.phylo
 #' @importFrom jsonlite fromJSON toJSON
-#' @importFrom openssl base64_decode
 #' @importFrom httr GET POST content stop_for_status upload_file warn_for_status
 #' add_headers timeout config
 #' @importFrom data.table rbindlist